Intentionally settting same value for min_sec and max_sec
Because when the scheduled livestream is late and off the schedule,
yt-dlp will wait for max_sec, which will likely to end up missing a lot of
livestream if the max_sec is long.

At first, I was setting max_sec to 3600 to avoid making too many request,
but it fails when the above mentioned is the case.